# Conflicts:
#	bootcamp/src/main/java/Spotify/controller/rest/model/AlbumRest.java
#	bootcamp/src/main/java/Spotify/mapper/AlbumMapper.java
#	bootcamp/src/main/java/Spotify/persistence/entity/AlbumEntity.java
#	bootcamp/src/test/java/Spotify/persistence/mapper/SongEntityMapperTest.java

+++ bootcamp/src/main/java/Spotify/controller/rest/model/ArtistRest.java
@@ -15,7 +15,7 @@ import java.util.List;
 public class ArtistRest implements Serializable {
 
     @JsonProperty("id")
-    private int id;
+    private Long id;
 
     @JsonProperty("name")
     private String name;
@@ -23,8 +23,6 @@ public class ArtistRest implements Serializable {
     @JsonProperty("description")
     private String description;
 
-    //@JsonProperty("albums")
-    //private List<AlbumRest> albums;
 
 
 }
